==> notebooks/NB01-DataCollection.ipynb <==
# Data Collection

## Install Neccesary Packages and Check for HTML Request

#!pip install requests scrapy pandas

import requests                # for sending HTTP requests
from scrapy import Selector    # for parsing HTML content
import pandas as pd            # for creating dataframe

from pprint import pprint

url = "https://editorial.rottentomatoes.com/guide/best-movies-of-all-time/"
html = requests.get(url).content  # This sends a GET request to the URL and stores the HTML content

sel = Selector(text=html) # instantiate

import requests                # for sending HTTP requests
from scrapy import Selector    # for parsing HTML content
import pandas as pd

## Create Functions to Gather Data from Website
We will be collecting data from Rotten Tomatoes' top 300 movie list. We first gather the ranking list, and then we find the individual descriptions within each movie by going into their specific pages

# from the ranking page, collect movie hyperlink, movie titile, and movie rank
def get_movie_links(url):
    html = requests.get(url).content
    sel = Selector(text=html)

    movie_links =sel.css('span.details a::attr(href)').getall()
    movie_titles = sel.css('span.details a::text').getall()
    movie_ranks = sel.css('td:nth-child(1)::text').getall()

    return {'link': movie_links,
            'title': movie_titles,
            'rank': movie_ranks}

url = 'https://editorial.rottentomatoes.com/guide/best-movies-of-all-time/'

#pprint(get_movie_links(url))

# convert the gathered data into a dataframe
top_df = pd.DataFrame(get_movie_links(url))
top_df

# go into each movie page and collect all information on it, from critic rating to box-office performance; if a piece of information is missing, fill it with none
def get_movie_info(url):
    html = requests.get(url).content
    sel = Selector(text=html)

    info_blocks = sel.css('div.content-wrap div.category-wrap')
    info_pairs = {}
    # there are some information which are hyperlinks and others that are pure text, so we have to operate on them differently
    links = ['Director', 'Producer', 'Screenwriter', 'Genre']
    texts = ['Distributor', 'Production Co', 'Rating', 'Original Language', 'Release Date (Theaters)', 'Release Date (Streaming)', 'Box Office (Gross USA)', 'Runtime', 'Sound Mix', 'Aspect Ratio']
    
    # collect description title and info from each text box
    for block in info_blocks:
        if block.css('dt.key rt-text::text').getall()[0] in links:
            info_pairs[f"{block.css('dt.key rt-text::text').getall()[0]}"] = block.css('dd rt-link::text').getall()
        elif block.css('dt.key rt-text::text').getall()[0] in texts:
            info_pairs[f"{block.css('dt.key rt-text::text').getall()[0]}"] = block.css('dd rt-text:not(rt-text.delimiter)::text').getall()
    
    # add rating scores
    info_pairs['critics_score'] = sel.css('rt-button:nth-child(3) > rt-text::text').getall()
    info_pairs['audience_score'] = sel.css('rt-button:nth-child(7) > rt-text::text').getall()

    # if a specific information was not represented on the webpage, change it to none for eventual dataframe
    for tag in links:
        if tag not in info_pairs.keys():
            info_pairs[tag] = None

    for tag in texts:
        if tag not in info_pairs.keys():
            info_pairs[tag] = None

    # reorder the columns to make them consistent for every movie page
    keys = list(info_pairs.keys())
    keys.sort()
    sorted_info_pairs = {i: info_pairs[i] for i in keys}

    return sorted_info_pairs

# test out the top movie page link (it has all datapoints filled)
get_movie_info('https://www.rottentomatoes.com/m/la_confidential')

# test out a movie link with some missing data, and we see it handles it alright
pprint(get_movie_info('https://www.rottentomatoes.com/m/saving_private_ryan'))

# get movie info for all pages in the top 300 list
top_df['info'] = top_df['link'].apply(get_movie_info)
top_df

# expand the dataframe to seperate each information into individual columns
top_df = pd.concat([top_df.drop(['info'], axis=1), top_df['info'].apply(pd.Series)], axis=1)
top_df

# save raw data into dataframe for data cleaning
top_df.to_json('../data/raw/movies.json')
