==> notebooks/NB02-Build-Database.ipynb <==
# Cleaning the Dataframe and Building a SQLite Database

## Install Neccesary Packages and Reading the JSON file

import pandas as pd
from datetime import datetime

top_df = pd.read_json("../data/raw/movies.json")

top_df.head() 

## Cleaning the Dataframe
We undo the list strucutre associated with many columns and convert them into the appropriate datatypes. Lastly, we format them in an ideal manner to create a database.

# though columns are represented as lists, some contain only one value; we create the columns as single objects
columns_to_explode = [
    'Original Language', 'Runtime', 'Release Date (Streaming)', 'Rating',
    'Release Date (Theaters)', 'Box Office (Gross USA)', 'audience_score', 'critics_score'
]

for column in columns_to_explode:
    top_df = top_df.explode(column)


# rename columns to more appropriate format and concise description
top_df = (
    top_df.rename(columns={
        'Box Office (Gross USA)': 'revenue',
        'Original Language': 'language',
        'Release Date (Streaming)': 'date-streaming',
        'Release Date (Theaters)': 'date-theater',
    })
)

top_df.head()

# clean rank column into integer
top_df['rank'] = top_df['rank'].astype('str')
top_df['rank'] = top_df['rank'].str.rstrip('.')
top_df['rank'] = top_df['rank'].apply(int)

def clean_revenue(revenue):
    revenue = revenue.replace('$','')
    
    # movie revenues are represented in millions, thousands, and billions
    quantifier = revenue[-1]

    if quantifier == 'M':
        return float(revenue[:-1])*1000000
    elif quantifier == 'K':
        return float(revenue[:-1])*1000
    elif quantifier == 'B':
        return float(revenue[:-1])*1000000000
    else:
        return float(revenue)


top_df['revenue'] = top_df['revenue'].apply(lambda x: clean_revenue(x) if x else None)

top_df.head()

def time_to_minutes(runtime):
    total_time = 0
    if 'h' and 'm' in runtime:
        halves = runtime.split()
        for half in halves:
            if 'h' in half:
                total_time += 60*int(half.replace('h',''))
            elif 'm' in half:
                total_time += int(half.replace('m',''))
        return total_time
    if 'm' in runtime:
        return runtime.rstrip('m')
        


top_df['Runtime'] = top_df['Runtime'].apply(time_to_minutes)

def clean_date(dateStr):
    dateStr = str(dateStr.split(',')[0])+str(dateStr.split(',')[1])

    date_format = '%b %d %Y'

    return datetime.strptime(dateStr, date_format).date()



top_df['date-theater'] = top_df['date-theater'].apply(lambda x: clean_date(x) if x else None)

top_df.head(10)

def clean_rating(ratingText):
    # seperate sub-ratings into individual elements within a new list
    if "(" in ratingText:
        words = ratingText.split("(")
        subRatings = words[1].rstrip(")").split("|")
        for subRating in subRatings:
            words.append(subRating)
        del words[1]
        words[0].strip()
        return words
    elif ratingText=="PG":
        return ["PG"]
    else:
        return list(ratingText)

top_df['Rating'] = top_df['Rating'].apply(lambda x: clean_rating(x) if x else None)

top_df.head()

# convert rating scores as int numbers
top_df['audience_score'] = top_df['audience_score'].apply(lambda x: x.replace('%',''))
top_df['critics_score'] = top_df['critics_score'].apply(lambda x: x.replace('%',''))
top_df['audience_score'] = top_df['audience_score'].astype(int)
top_df['critics_score'] = top_df['critics_score'].astype(int)

top_df.head()

# this function converts lists into strings where each element is seperated by "%%%"
def list_to_string_converter(list_from_df):
    if not isinstance(list_from_df, list):
        return ""
    final_str = ""
    for items in list_from_df:
        final_str = final_str + "%%%" + str(items)
    return final_str

# we apply this function to create a sql database, as it cannot handle lists within columns
columns_to_convert = ['Aspect Ratio', 'Director', 'Distributor', 'Genre', 
                      'Producer', 'Production Co', 'Rating', 'Screenwriter', 'Sound Mix']
for column in columns_to_convert:
    top_df[column] = top_df[column].apply(lambda x: list_to_string_converter(x) if isinstance(x, list) else "")

top_df

# every column is of the datatype we want
top_df.info()

## Creating the SQL Database
We create three tables, one for general movie information, another for technical details, another for reception-related statistics. In the end, we pd merge them together.

# packages for sql database creation
import os
import sqlite3

DATA_FOLDER = os.path.join('../data/clean/')

conn = sqlite3.connect(os.path.join(DATA_FOLDER, 'moviedatabase.db'))

selected_columns = ['link', 'title', 'Director', 'Genre', 'language', 'date-streaming', 'date-theater', 
                    'Runtime', "Rating"]

movies = top_df[selected_columns].copy().drop_duplicates()

movies.to_sql('movies', conn, if_exists='replace', index=False)

pd.read_sql('SELECT * FROM movies LIMIT 5', conn)


technicalities = (
    top_df[['link', 'Aspect Ratio', 'Distributor', 'Sound Mix','Producer', 'Production Co', 'Screenwriter']].drop_duplicates()
)

technicalities.to_sql('technicalities', conn, if_exists='replace', index=False)

post_release = (
    top_df[['link','rank', 'revenue', 'audience_score', 'critics_score']].drop_duplicates()
)

post_release.to_sql('post_release', conn, if_exists='replace', index=False)

pd.read_sql('SELECT * FROM post_release LIMIT 5', conn)


(
    pd.read_sql('SELECT * FROM movies', conn)
        .merge(pd.read_sql('SELECT * FROM technicalities', conn), 
                left_on='link', 
                right_on='link', 
                how='left')
        .merge(pd.read_sql('SELECT * FROM post_release', conn), 
                left_on='link', 
                right_on='link', 
                how='left')
)
