==> notebooks/NB03-EDA-and-Dataviz.ipynb <==
# Data Analysis and Visualizations

## Import Packages and Read in Database as Pandas DataFrame
We take the cleaned database and read it as a dataframe by merging our three tables together on the link variable. We also expand on columns which originally contained lists by seperating the strings by our delimiter "%%%".

import os
import sqlite3
import pandas as pd

DATA_FOLDER = os.path.join('../data/clean/')
conn = sqlite3.connect(os.path.join(DATA_FOLDER, 'moviedatabase.db'))

top_df = (
    pd.read_sql('SELECT * FROM movies', conn)
        .merge(pd.read_sql('SELECT * FROM technicalities', conn), 
                left_on='link', 
                right_on='link', 
                how='left')
        .merge(pd.read_sql('SELECT * FROM post_release', conn), 
                left_on='link', 
                right_on='link', 
                how='left')
)

def crypted_to_list(list_value):
    return list_value.split("%%%")[1:]

columns_to_convert = ['Aspect Ratio', 'Director', 'Distributor', 'Genre', 
                      'Producer', 'Production Co', 'Rating', 'Screenwriter', 'Sound Mix']
for column in columns_to_convert:
    top_df[column] = top_df[column].apply(lambda x: crypted_to_list(x) if isinstance(x, str) else [])

top_df.head()

## Early Data Analysis
Play around and see any interesting trends or pieces of information

# Top directors here make sense, as each are notable figures within the film industry
top_df.explode('Director')['Director'].value_counts()

# Andrew Stanton and Pete Docter both worked on Pixar films; Akira Kurosawa directed and wrote his screenplays
top_df.explode('Screenwriter')['Screenwriter'].value_counts()

# We see that there are films with multiple genres, with drama, comedy, adventure being the most popular
top_df.explode('Genre')['Genre'].value_counts()

# make revenues in millions to read numbers easier
top_df['revenue'] = top_df['revenue'].apply(lambda x: x/1000000)
top_df.explode('Genre').groupby('Genre')['revenue'].describe().sort_values(by=['mean'])

top_df['critics_score'].describe()

top_df['audience_score'].describe()

Interestingly, we see that the audience scores have a higher standard deviation than the critics showing that critics have a more homogenous opinion than the audience. The fact that the mean is lower for the audience means that appreciate these movies less than the critics. 

genre_critics = top_df.explode('Genre').groupby('Genre')['critics_score'].describe().sort_values(by=['mean'])
genre_critics = genre_critics.drop(genre_critics[genre_critics['count'] < 10].index)
genre_critics.sort_values(by=['mean'])

genre_audience = top_df.explode('Genre').groupby('Genre')['audience_score'].describe().sort_values(by=['mean'])
genre_audience = genre_audience.drop(genre_audience[genre_audience['count'] < 10].index)
genre_audience.sort_values(by=['mean'])

Among genres, critics and the audience have noticeably different opinions. This is a point of interest I will explore later.

# Top grossing movies are all part of large franchises
top_df.sort_values(by=['revenue'], ascending=False).head()

# The two Star Wars movies seem to have incorrect revenue values as they are $123 and $345. Besides that, it's interesting to see "All About Eve", whiched ranked no 16 on the list, appear in the bottom 5 here
top_df.sort_values(by=['revenue']).head()

# Run time of movies vary significantly and the longest movie is six times as long as the shortest.
top_df['Runtime'].describe()

After collecting these initial impressions about the top movies lists, I want to explore the interplay between critic and audience scores, see what types of movies succeed, and how language and genre can have an affect on the reception of movies.

## Data Visualizations

# For plotting purposes
import math 
import cairosvg
from lets_plot import *
LetsPlot.setup_html()

## Analyzing the Interplay Between Critics Score and Audience Score

# Create new df to show genre behavior among films
# Also put log scale on revenue
# Interestingly, both Star Wars movie below have registered revenues in the US as $123 and $345
# As misinputed values, I remove
genre_df = top_df.explode('Genre')
genre_df['revenue'] = genre_df['revenue'].apply(lambda x: math.log10(x))
genre_df = genre_df.drop(genre_df[genre_df.title == "Star Wars: The Force Awakens"].index)
genre_df = genre_df.drop(genre_df[genre_df.title == "Star Wars: Episode IV - A New Hope"].index)


# Revenue vs audience score scatterplot
tooltip_setup1 = (
    layer_tooltips()
        .line('@title')
)

scatter_plot1 = (
    ggplot(genre_df, aes(x='revenue', y='audience_score', color='Genre')) +
    geom_jitter(tooltips=tooltip_setup1) +
    scale_x_continuous(name='Revenue') +
    scale_y_continuous(name='Audience Score', limits=[75, 100]) +
    labs(title='There Seems to be a Slight Correlation Between Revenue and Audience Score',
         x='Log 10 Revenue in Millions',
         y='Audience Score') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(750, 600)
)

scatter_plot1
ggsave(scatter_plot1, 'revenue_audience.png', path='../docs/figures')

# Revenue vs critics score scatter plot
scatter_plot2 = (
    ggplot(genre_df, aes(x='revenue', y='critics_score', color='Genre')) +
    geom_jitter(tooltips=tooltip_setup1) +
    scale_x_continuous(name='Revenue') +
    scale_y_continuous(name='Critics Score', limits=[90, 100]) +
    labs(title='There is No Correlation Between Revenue and Critics Score',
         x='Log 10 Revenue in Millions',
         y='Critics Score') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(750,600)
)

scatter_plot2
ggsave(scatter_plot2, 'revenue_critics.png', path='../docs/figures')

# Audience vs Critics Score
scatter_plot3 = (
    ggplot(top_df.explode('Genre'), aes(x='audience_score', y='critics_score', color='Genre')) +
    geom_jitter(tooltips=tooltip_setup1) +
    scale_x_continuous(name='Audience Score', limits=[75, 100]) +
    scale_y_continuous(name='Critics Score', limits=[90, 100]) +
    labs(title='There Seems to be no Linear Relationship',
         x='Audience Score',
         y='Critics Score') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    )
    + ggsize(750,600)
)

scatter_plot3
ggsave(scatter_plot3, 'audience_critics.png', path='../docs/figures')

# Find difference in mean rating between critics and audience within different genres
diff_df = genre_critics['mean'] - genre_audience['mean']
diff_df = diff_df.to_frame().reset_index().sort_values(by=['mean'])

scatter_plot4 = (
    ggplot(diff_df, aes(x='Genre', y='mean',)) +
    geom_point() +
    scale_x_continuous(name='Genre') +
    scale_y_continuous(name='Rating Difference') +
    labs(title='Musicals are Underpreciated among the Audience and War Films are almost Equally liked',
         x='Genre',
         y='Rating Differnce (Critics Mean - Audience Mean Score)') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(800,600)
)

scatter_plot4
ggsave(scatter_plot4, 'diff_mean_genre.png', path='../docs/figures')

## A Brief Look at Language, Release Data, and Genre

# Plot rank vs revenue with categories on language
tooltip_setup2 = (
    layer_tooltips()
        .line('@title')
        .line('@language')
        .line('$@revenue m')
)

scatter_plot5 = (
    ggplot(top_df, aes(x='rank', y='revenue', color='language')) +
    geom_jitter(tooltips=tooltip_setup2) +
    scale_x_continuous(name='Rank') +
    scale_y_continuous(name='Revenue') +
    labs(title='Better Rank Does Not Indicate More Revenue',
         x='Rank',
         y='Revenue') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(750,500)
)

scatter_plot5
ggsave(scatter_plot5, 'rank_revenue.png', path='../docs/figures')

# Rank vs runtime by language, excluding English as it overpopulates the list
language_df = top_df.drop(top_df[top_df.language == "English"].index)

tooltip_setup3 = (
    layer_tooltips()
        .line('@title')
        .line('@language')
)

scatter_plot6 = (
    ggplot(language_df, aes(x='rank', y='Runtime', color='language')) +
    geom_jitter(tooltips=tooltip_setup3) +
    scale_x_continuous(name='Rank') +
    scale_y_continuous(name='Run Time') +
    labs(title='Run Time Does Not Indicate Rank Nor is There a Clear Pattern with Languages',
         x='Rank',
         y='Run Time') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(750,500)
)

scatter_plot6
ggsave(scatter_plot6, 'rank_runtime.png', path='../docs/figures')

# Plot date vs rank
date_df = top_df.sort_values(by=['date-theater'])

scatter_plot7 = (
    ggplot(date_df, aes(x='date-theater', y='rank', color='language')) +
    geom_point(tooltips=tooltip_setup1) +
    scale_x_datetime() +
    scale_y_continuous(name='Rank') +
    labs(title='Release Date has no Influence on Rank',
         x='Theater Release Date',
         y='Rank') +
    theme(
        axis_text_x=element_text(size=12),
        axis_text_y=element_text(size=12),
        axis_title_x=element_text(size=14),
        axis_title_y=element_text(size=14),
        plot_title=element_text(size=16, face='bold')
        
    ) +
    ggsize(750,500)
)

scatter_plot7
ggsave(scatter_plot7, 'date_rank.png', path='../docs/figures')

# Plot distribution of runtime of movies based on genres

# Drop genres with less than 10 movies
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Music"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Anime"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Sports"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Documentary"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "LGBTQ+"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Western"].index)
genre_df = genre_df.drop(genre_df[genre_df.Genre == "Holiday"].index)


p1 = (
    ggplot(genre_df, aes(x="Runtime", color="Genre")) +
    ggsize(1000, 500) +
    geom_density() + 
    labs(title='Distribution of Runtimes by Genre',
         subtitle="War Films are the Longest with Comedy being the Shortest") 
         )
p1
ggsave(p1, 'runtime_distribution_genre.png', path='../docs/figures')
